+++ R/get_insee.R
@@ -5,9 +5,11 @@
 #' It is used by the functions get_insee_dataset, get_insee_idbank and get_dataset_list.
 #' The data is cached, hence all queries are only run once per R session.
 #' The user can disable the download display in the console with the following command :
-#' Sys.setenv(INSEE_download_verbose = "FALSE")
-#' The use of cached data can be disabled with : Sys.setenv(INSEE_no_cache_use = "TRUE")
+#' Sys.setenv(INSEE_download_verbose = "FALSE").
+#' The use of cached data can be disabled with : Sys.setenv(INSEE_no_cache_use = "TRUE").
 #' All queries are printed in the console with this command: Sys.setenv(INSEE_print_query = "TRUE").
+#' The RapidXML C++ library is used as a backup thanks to the readsdmx package.
+#' It can be used instead of the internal parser with this command : Sys.setenv(INSEE_read_sdmx_fast = "TRUE")
 #' @param link SDMX query link
 #' @param step argument used only for internal package purposes to tweak download display
 #' @return a tibble containing the data
@@ -29,6 +31,7 @@ get_insee = function(link, step = "1/1"){
   insee_value_as_numeric = if(Sys.getenv("INSEE_value_as_numeric") == "TRUE"){TRUE}else{FALSE}
   insee_print_query = if(Sys.getenv("INSEE_print_query") == "TRUE"){TRUE}else{FALSE}
   insee_no_cache_use = if(Sys.getenv("INSEE_no_cache_use") == "TRUE"){TRUE}else{FALSE}
+  insee_read_sdmx_fast = if(Sys.getenv("INSEE_read_sdmx_fast") == "TRUE"){TRUE}else{FALSE}
 
   if(insee_download_verbose){
     if(insee_print_query == TRUE) {
@@ -41,7 +44,20 @@ get_insee = function(link, step = "1/1"){
 
   if((!file.exists(file_cache)) | insee_no_cache_use){
 
-    data_final = read_sdmx_slow(link, step)
+    # by default the internal parser is used, if it fails the readsdmx parser is used
+    if(!insee_read_sdmx_fast){
+      data_final = read_sdmx_slow(link, step)
+
+      if(is.null(data_final)){
+        data_final = read_sdmx_fast(link, step)
+      }
+    }else{
+      data_final = read_sdmx_fast(link, step)
+
+      if(is.null(data_final)){
+        data_final = read_sdmx_slow(link, step)
+      }
+    }
 
     if(!is.null(data_final)){
 
@@ -52,6 +68,9 @@ get_insee = function(link, step = "1/1"){
 
         message(crayon::style(msg, "green"))
       }
+    }else{
+      msg = "An error occurred"
+      message(crayon::style(msg, "red"))
     }
   }else{